front: move useDebounce and useDebouncedFunc in hooks folder
