+++ front/src/applications/stdcm/components/StdcmForm/StopDurationInput.tsx
@@ -7,7 +7,7 @@ import { updateStdcmPathStep } from 'reducers/osrdconf/stdcmConf';
 import type { StdcmPathStep } from 'reducers/osrdconf/types';
 import { useAppDispatch } from 'store';
 import { Duration } from 'utils/duration';
-import { useDebounce } from 'utils/helpers';
+import { useDebounce } from 'utils/hooks/useDebounce';
 import { parseNumber } from 'utils/strings';
 
 import { StdcmStopTypes } from '../../types';
